backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.3
psycopg2==2.8.5
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
uWSGI==2.0.18
